Add option to print words
Now, if you run the script as normal, it will print out the phrases from the
random beginning to the end goal. It will only show the chosen child from each
litter, so it looks like a linear evolution. If you don't want to see this, just
do an average with 1 trial.